pyproject.toml
[project]
name = "django-assignment"
version = "0.1.0"
description = ""
authors = [
    {name = "rodzlen",email = "qkqhrkxdkdy@naver.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-extensions (>=3.2.3,<4.0.0)",
    "ipython (>=8.31.0,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
